# docs/swagger.yaml
openapi: 3.0.3
info:
  title: API Gateway Service
  description: |
    Multi-channel messaging API Gateway that provides unified access to SMS, Email, and WhatsApp messaging services.
    
    ## Features
    - Multi-channel message sending (SMS, Email, WhatsApp)
    - Bulk message processing
    - Real-time webhook handling
    - Rate limiting and authentication
    - Conversation management
    - Channel health monitoring
    
    ## Authentication
    Supports two authentication methods:
    - **JWT Bearer Token**: For user-based authentication
    - **API Key**: For service-to-service authentication (X-API-Key header)
    
    ## Rate Limiting
    Hierarchical rate limiting is applied:
    - Global: 1000 requests/second
    - Per-tenant: 50 requests/second
    - Per-channel: Channel-specific limits
    
  version: 1.0.0
  contact:
    name: API Gateway Team
    email: api-gateway@company.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api-gateway.company.com/api/v1
    description: Production server
  - url: https://staging-api-gateway.company.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /messages/{messageId}:
    get:
      tags:
        - Messages
      summary: Get message status
      description: Retrieve the current status and details of a specific message
      operationId: getMessageStatus
      parameters:
        - name: messageId
          in: path
          required: true
          description: Unique message identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Message details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageStatus'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /channels/{channelId}:
    get:
      tags:
        - Channels
      summary: Get channel details
      description: Get detailed information about a specific messaging channel
      operationId: getChannelDetails
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            enum: [sms, email, whatsapp]
      responses:
        '200':
          description: Channel details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelDetails'
        '404':
          $ref: '#/components/responses/NotFound'
  /channels/{channelId}/status:
    get:
      tags:
        - Channels
      summary: Get channel health status
      description: Check the current health and operational status of a messaging channel
      operationId: getChannelStatus
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            enum: [sms, email, whatsapp]
      responses:
        '200':
          description: Channel status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelStatus'
  /channels/{channelId}/config:
    patch:
      tags:
        - Channels
      summary: Update channel configuration
      description: Update configuration settings for a specific channel
      operationId: updateChannelConfig
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            enum: [sms, email, whatsapp]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelConfigUpdate'
      responses:
        '200':
          description: Channel configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelConfig'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
  /channels/{channelId}/test:
    post:
      tags:
        - Channels
      summary: Test channel
      description: Send a test message through the specified channel to verify configuration
      operationId: testChannel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            enum: [sms, email, whatsapp]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelTestRequest'
      responses:
        '200':
          description: Test message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelTestResponse'

  # Conversation APIs
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service authentication
  schemas:
...
    MessageStatus:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            messageId:
              type: string
              format: uuid
            status:
              type: string
              enum: [queued, sent, delivered, failed, expired]
            channel:
              type: string
            recipient:
              type: string
            createdAt:
              type: string
              format: date-time
            sentAt:
              type: string
              format: date-time
            deliveredAt:
              type: string
              format: date-time
            failedAt:
              type: string
              format: date-time
            errorCode:
              type: string
            errorMessage:
              type: string
            retryCount:
              type: integer
            metadata:
              type: object
...
    ChannelDetails:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            status:
              type: string
              enum: [active, inactive, maintenance]
            configuration:
              type: object
            capabilities:
              type: array
              items:
                type: string
            rateLimits:
              type: object
              properties:
                perSecond:
                  type: integer
                perMinute:
                  type: integer
                perDay:
                  type: integer
    ChannelStatus:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            channel:
              type: string
            status:
              type: string
              enum: [healthy, degraded, down]
            lastChecked:
              type: string
              format: date-time
            responseTime:
              type: integer
              description: Response time in milliseconds
            errorRate:
              type: number
              description: Error rate percentage
            uptime:
              type: number
              description: Uptime percentage
    ChannelConfigUpdate:
      type: object
      properties:
        webhookUrl:
          type: string
          format: uri
        webhookSecret:
          type: string
          minLength: 32
        rateLimits:
          type: object
          properties:
            perSecond:
              type: integer
              minimum: 1
              maximum: 100
            perMinute:
              type: integer
              minimum: 1
              maximum: 5000
            perDay:
              type: integer
              minimum: 1
              maximum: 100000
        settings:
          type: object
          additionalProperties: true
    ChannelConfig:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            channel:
              type: string
            configuration:
              type: object
            updatedAt:
              type: string
              format: date-time
    ChannelTestRequest:
      type: object
      properties:
        recipient:
          type: string
          description: Test recipient (optional, uses default if not provided)
        message:
          type: string
          maxLength: 1000
          description: Custom test message
        testType:
          type: string
          enum: [send, webhook, both]
          default: send
    ChannelTestResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            testId:
              type: string
            results:
              type: object
              properties:
                send:
                  type: object
                  properties:
                    success:
                      type: boolean
                    messageId:
                      type: string
                    error:
                      type: string
                webhook:
                  type: object
                  properties:
                    success:
                      type: boolean
                    responseTime:
                      type: integer
                    error:
                      type: string

    # Conversation Schemas
...
    ResponseMeta:
      type: object
      properties:
        requestId:
          type: string
          description: Unique request identifier for tracing
        timestamp:
          type: string
          format: date-time
        version:
          type: string
          description: API version
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
        meta:
          $ref: '#/components/schemas/ResponseMeta'
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: VALIDATION_ERROR
                      message:
                        example: Request validation failed
                      details:
                        type: array
                        items:
                          type: object
                          properties:
                            field:
                              type: string
                            message:
                              type: string
                            value:
                              type: string
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: UNAUTHORIZED
                      message:
                        example: Invalid or missing authentication token
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: NOT_FOUND
                      message:
                        example: The requested resource was not found
...